+++ api/src/certification/enrolment/infrastructure/repositories/index.js
@@ -9,6 +9,7 @@ import * as certificationCpfCityRepository from './certification-cpf-city-reposi
 import * as certificationCpfCountryRepository from './certification-cpf-country-repository.js';
 import * as complementaryCertificationRepository from './complementary-certification-repository.js';
 import * as countryRepository from './country-repository.js';
+import * as scoCertificationCandidateRepository from './sco-certification-candidate-repository.js';
 import * as sessionForAttendanceSheetRepository from './session-for-attendance-sheet-repository.js';
 import * as sessionRepository from './session-repository.js';
 
@@ -26,19 +27,21 @@ import * as sessionRepository from './session-repository.js';
  * @typedef {sessionForAttendanceSheetRepository} SessionForAttendanceSheetRepository
  * @typedef {sessionManagementRepository} SessionManagementRepository
  * @typedef {countryRepository} CountryRepository
+ * @typedef {scoCertificationCandidateRepository} ScoCertificationCandidateRepository
  */
 const repositoriesWithoutInjectedDependencies = {
   candidateRepository,
   centerRepository,
-  complementaryCertificationRepository,
-  sessionRepository,
   certificationCandidateRepository,
   certificationCenterRepository,
   certificationCpfCountryRepository,
   certificationCpfCityRepository,
+  complementaryCertificationRepository,
+  countryRepository,
+  scoCertificationCandidateRepository,
   sessionForAttendanceSheetRepository,
   sessionManagementRepository,
-  countryRepository,
+  sessionRepository,
 };
 
 /**
